Fyyur/requirements.txt
babel
python-dateutil==2.6.0
flask-moment
flask-wtf
flask-script
flask-migrate
psycopg2